+++ repository/event_repository.go
@@ -3,11 +3,11 @@ package repository
 import "github.com/SaloEater/WhatNot-Webhook-Holder/entity"
 
 type EventRepository interface {
-	GetAllByBreak(int) ([]*entity.Event, error)
-	Get(int) (*entity.Event, error)
-	Delete(int) error
+	GetAllByBreak(int64) ([]*entity.Event, error)
+	Get(int64) (*entity.Event, error)
+	Delete(int64) error
 	Update(*entity.Event) error
-	Create(*entity.Event) (int, error)
-	GetAllChildren(int) ([]*entity.Event, error)
+	Create(*entity.Event) (int64, error)
+	GetAllChildren(int64) ([]*entity.Event, error)
 	UpdateAll([]*entity.Event) error
 }
