+++ service/break_delete.go
@@ -1,7 +1,7 @@
 package service
 
 type DeleteBreakRequest struct {
-	Id int `json:"id"`
+	Id int64 `json:"id"`
 }
 
 type DeleteBreakResponse struct {
